backend/requirements.txt
Cerberus==1.3.3
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Cors==3.0.10
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==2.1.0
pymongo==3.12.0
six==1.16.0
typing-extensions==3.10.0.2
uuid==1.30
Werkzeug==2.0.1
zipp==3.5.0